backend/services/knowledge_base/requirements.txt
fastapi==0.115.12
openai==1.67.0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.5
qdrant_client==1.13.2
uvicorn==0.34.0